mlops_pipeline/requirements.txt
requests
numpy<2.0
mlflow
scikit-learn
pandas
tqdm
Pillow
torch
torchvision
cupy